3rd commit

+++ services/src/main/java/com/news/service/user/UserService.java
@@ -12,28 +12,46 @@ import com.news.repository.UserSearchRepo;
 
 @Service
 public class UserService {
-	
+
 	@Autowired
 	private UserRepo userRepo;
-	
+
 	@Autowired
 	private UserSearchRepo userSearchRepo;
-	
-	public String searchSave(Search search) {
-		search.setUser(userRepo.findByUserEmail(search.getUser().getUserEmail()));
-		userSearchRepo.save(search);
-		return "saved";
-	}
-	
-	public List<Search> searchHistory(String userEmail) {
-		User user = userRepo.findByUserEmail(userEmail);
-		return userSearchRepo.findDetailsByUser(user);
 
+	public String searchSave(Search search) {
+		if (search.getUser().getUserEmail() != null) {
+			if (search.getSearchString() != null) {
+				if (userRepo.findByUserEmail(search.getUser().getUserEmail()) != null) {
+					search.setUser(userRepo.findByUserEmail(search.getUser().getUserEmail()));
+					userSearchRepo.save(search);
+					return "saved";
+				} 
+				else
+					return "no such user present";
+			} 
+			else
+				return "no search data present";
+		} 
+		else
+			return "no email is present";
 	}
+
+	public List<Search> searchHistory(String userEmail) {
+		if (userRepo.findByUserEmail(userEmail) != null){
+		User user = userRepo.findByUserEmail(userEmail);
+			return userSearchRepo.findDetailsByUser(user);
+		}
+		else
+			return null;
+	}
+
 	public String searchDelete(long searchId) {
-		userSearchRepo.deleteById(searchId);
-		return "deleted";
+		if (userSearchRepo.findById(searchId).isPresent()) {
+			userSearchRepo.deleteById(searchId);
+			return "deleted";
+		} else
+			return "Id does not exist";
 	}
-	
-	
+
 }
